Updated demo scripts for Openshift

+++ k8s-ubuntu/illumio-demo-apps/illumio-cven-install.sh
@@ -1,1 +1,8 @@
-helm upgrade --install illumio -f illumio-cven.yaml oci://quay.io/illumio/illumio --namespace illumio-system --create-namespace
+read -r -p "Is this an OpenShift cluster? [y/N] " openshift_answer
+openshift_answer="${openshift_answer:-N}"
+
+if [[ "$openshift_answer" =~ ^[Yy]$ ]]; then
+  helm upgrade --install illumio -f illumio-cven.yaml oci://quay.io/illumio/illumio --namespace illumio-system --create-namespace --server-side=false
+else
+  helm upgrade --install illumio -f illumio-cven.yaml oci://quay.io/illumio/illumio --namespace illumio-system --create-namespace
+fi
